# src/openapi.json
{
  "swagger": "2.0",
  "info": {
    "title": "FurWaz API Documentation",
    "description": "Documentation for the FurWaz API",
    "version": "1.0.0",
    "contact": {
      "name": "FurWaz",
      "url": "https://furwaz.fr",
      "email": "contact@furwaz.fr"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "authorization"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/portal/generate": {
      "get": {
        "tags": [
          "Portal"
        ],
        "description": "Generate a new portal token",
        "operationId": "generatePortalToken",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/portal/{portal}/user": {
      "get": {
        "tags": [
          "Portal"
        ],
        "description": "Get the portal user",
        "operationId": "getPortalUser",
        "parameters": [
          {
            "name": "portal",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/portal/{portal}/app": {
      "get": {
        "tags": [
          "Portal"
        ],
        "description": "Get the portal app",
        "operationId": "getPortalApp",
        "parameters": [
          {
            "name": "portal",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/portal/{portal}": {
      "post": {
        "tags": [
          "Portal"
        ],
        "description": "Connect a user to the portal",
        "operationId": "connectPortalUser",
        "parameters": [
          {
            "name": "portal",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/store/cart/product/{productId}": {
      "get": {
        "tags": [
          "Cart"
        ],
        "description": "Get cart product",
        "operationId": "getUserProduct",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Cart"
        ],
        "description": "Update cart product",
        "operationId": "updateUserProduct",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Cart"
        ],
        "description": "Delete cart product",
        "operationId": "deleteUserProduct",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/store/product/": {
      "post": {
        "tags": [
          "Product"
        ],
        "description": "Create a new product",
        "operationId": "createProduct",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "price": {
                  "example": "any"
                },
                "app": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "Product"
        ],
        "description": "Get own app products",
        "operationId": "getOwnAppProducts",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/store/product/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "description": "Get a product",
        "operationId": "getProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Product"
        ],
        "description": "Update a product",
        "operationId": "updateProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "price": {
                  "example": "any"
                },
                "app": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "description": "Delete a product",
        "operationId": "deleteProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "price": {
                  "example": "any"
                },
                "app": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/store/checkout/": {
      "post": {
        "tags": [
          "Cart"
        ],
        "description": "Create a new product",
        "operationId": "createProduct",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "Cart"
        ],
        "description": "Get own app products",
        "operationId": "getOwnAppProducts",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
  },
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization"
      }
    }
  }
}